+++ mood-backend/src/config.ts
@@ -11,5 +11,12 @@ const defaults = {
   PORT: 3000,
 };
 
+export const spotifyConfig = {
+  authUrl: 'https://accounts.spotify.com/api/token',
+  baseUrl: 'https://api.spotify.com/v1',
+  clientId: process.env.SPOTIFY_CLIENT_ID || '',
+  clientSecret: process.env.SPOTIFY_CLIENT_SECRET || '',
+};
+
 export const environment = (process.env.NODE_ENV as Environment) ?? defaults.ENVIRONMENT;
 export const port = process.env.PORT ?? defaults.PORT;
